refactor: 공통으로 사용되는 클래스 common 디렉토리로 이동
