+++ server/module-api/src/test/java/com/momo/common/RestDocsControllerTest.java
@@ -1,4 +1,4 @@
-package com.momo;
+package com.momo.common;
 
 import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.documentationConfiguration;
 import static org.springframework.restdocs.operation.preprocess.Preprocessors.prettyPrint;
