+++ server/module-api/src/test/java/com/momo/district/docs/DistrictRestDocsTest.java
@@ -6,7 +6,7 @@ import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuild
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-import com.momo.RestDocsControllerTest;
+import com.momo.common.RestDocsControllerTest;
 import com.momo.api.district.DistrictController;
 import com.momo.domain.district.entity.District;
 import com.momo.domain.district.repository.DistrictRepository;
