+++ server/module-api/src/test/java/com/momo/group/docs/GroupRestDocsTest.java
@@ -1,8 +1,8 @@
 package com.momo.group.docs;
 
 
-import static com.momo.CommonFileUploadSupport.uploadMockSupport;
-import static com.momo.CommonFileUploadSupport.uploadTestFile;
+import static com.momo.common.CommonFileUploadSupport.uploadMockSupport;
+import static com.momo.common.CommonFileUploadSupport.uploadTestFile;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.when;
 import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders.fileUpload;
@@ -11,7 +11,7 @@ import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuild
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-import com.momo.RestDocsControllerTest;
+import com.momo.common.RestDocsControllerTest;
 import com.momo.api.group.GroupController;
 import com.momo.domain.district.entity.City;
 import com.momo.domain.group.dto.GroupCreateRequest;
