+++ server/module-api/src/test/java/com/momo/group/docs/GroupsRestDocsTest.java
@@ -7,7 +7,7 @@ import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuild
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-import com.momo.RestDocsControllerTest;
+import com.momo.common.RestDocsControllerTest;
 import com.momo.api.group.GroupsController;
 import com.momo.domain.district.entity.City;
 import com.momo.domain.group.dto.GroupCardResponse;
