+++ server/module-api/src/test/java/com/momo/post/acceptance/step/PostAcceptanceStep.java
@@ -1,6 +1,6 @@
 package com.momo.post.acceptance.step;
 
-import static com.momo.CommonFileUploadSupport.uploadAssuredSupport;
+import static com.momo.common.CommonFileUploadSupport.uploadAssuredSupport;
 import static io.restassured.RestAssured.given;
 import static org.assertj.core.api.Assertions.assertThat;
