+++ server/module-api/src/test/java/com/momo/schedule/docs/ScheduleRestDocsTest.java
@@ -8,7 +8,7 @@ import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuild
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-import com.momo.RestDocsControllerTest;
+import com.momo.common.RestDocsControllerTest;
 import com.momo.api.schedule.ScheduleController;
 import com.momo.domain.common.dto.EnumResponse;
 import com.momo.domain.group.entity.Category;
